refactor(form-builder): move calendar components to separate files
