+++ packages/@sanity/form-builder/src/inputs/DateInputs/base/calendar/CalendarMonth.tsx
@@ -1,0 +1,52 @@
+import {Box, Grid, Text} from '@sanity/ui'
+import {isSameDay, isSameMonth} from 'date-fns'
+import React from 'react'
+import {CalendarDay} from './CalendarDay'
+import {WEEK_DAY_NAMES} from './constants'
+import {getWeeksOfMonth} from './utils'
+
+interface CalendarMonthProps {
+  date: Date
+  focused?: Date
+  selected?: Date
+  onSelect: (date: Date) => void
+  hidden?: boolean
+}
+
+export function CalendarMonth(props: CalendarMonthProps) {
+  return (
+    <Box aria-hidden={props.hidden || false}>
+      <Grid columns={7} gap={1}>
+        {WEEK_DAY_NAMES.map((weekday) => (
+          <Box key={weekday} paddingY={1}>
+            <Text size={1} weight="medium" style={{textAlign: 'center'}}>
+              {weekday}
+            </Text>
+          </Box>
+        ))}
+
+        {getWeeksOfMonth(props.date).map((week, weekIdx) =>
+          week.days.map((date, dayIdx) => {
+            const focused = props.focused && isSameDay(date, props.focused)
+            const selected = props.selected && isSameDay(date, props.selected)
+            const isToday = isSameDay(date, new Date())
+            const isCurrentMonth = props.focused && isSameMonth(date, props.focused)
+
+            return (
+              <CalendarDay
+                date={date}
+                focused={focused}
+                isCurrentMonth={isCurrentMonth}
+                isToday={isToday}
+                // eslint-disable-next-line react/no-array-index-key
+                key={`${weekIdx}-${dayIdx}`}
+                onSelect={props.onSelect}
+                selected={selected}
+              />
+            )
+          })
+        )}
+      </Grid>
+    </Box>
+  )
+}
