+++ packages/@sanity/form-builder/src/inputs/DateInputs/base/calendar/YearInput.tsx
@@ -1,6 +1,6 @@
 import React from 'react'
 import {TextInput} from '@sanity/ui'
-import {LazyTextInput} from './LazyTextInput'
+import {LazyTextInput} from '../LazyTextInput'
 
 type Props = Omit<React.ComponentProps<typeof TextInput>, 'onChange' | 'value'> & {
   value?: number
